+++ src/antity-pgsql.ts
@@ -1,2 +1,3 @@
 export { SQLEntity } from './entity';
 export { filter } from './filter/filter';
+export { execute } from './crud/execute';