+++ src/crud/insert.ts
@@ -1,6 +1,7 @@
-import { arrayAdd } from "@dwtechs/sparray";
+import { add as spAdd } from "@dwtechs/sparray";
 import { execute as exe  } from "./execute";
 import { $i } from "./i";
+import { quoteIfUppercase } from "./quote";
 import type { PGResponse, Filter } from "../types";
 
 export class Insert {
@@ -10,8 +11,7 @@ export class Insert {
   private _nbProps: number = 2;
 
   public addProp(prop: string): void {
-    this._props = arrayAdd(prop, this._props.length - 2);
-    // this._props.splice(this._props.length - 2, 0, prop);
+    this._props = spAdd(this._props, quoteIfUppercase(prop), this._props.length - 2) as string[];
     this._cols = this._props.join(", ");
     this._nbProps++;
   }
@@ -34,7 +34,7 @@ export class Insert {
     consumerName: string,
     rtn: string = "",
   ): { query: string, args: (Filter["value"])[] } {
-    let query = `INSERT INTO "${table}" (${this._cols}) VALUES `;
+    let query = `INSERT INTO ${quoteIfUppercase(table)} (${this._cols}) VALUES `;
     const args: (Filter["value"])[] = [];
     let i = 0;
     for (const row of rows) {