+++ src/crud/update.ts
@@ -1,14 +1,14 @@
-import { arrayAdd } from "@dwtechs/sparray";
+import { add as spAdd } from "@dwtechs/sparray";
 import { execute as exe } from "./execute";
 import { $i } from "./i";
+import { quoteIfUppercase } from "./quote";
 import type { PGResponse, Filter } from "../types";
 
 export class Update {
   private _props: string[] = ["consumerId", "consumerName"];
 
   public addProp(prop: string): void {
-    this._props = arrayAdd(prop, this._props.length - 2);
-    // this._props.splice(this._props.length - 2, 0, prop);
+    this._props = spAdd(this._props, quoteIfUppercase(prop), this._props.length - 2) as string[];
   }
 
   public query(
@@ -20,7 +20,7 @@ export class Update {
     rows = this.addConsumer(rows, consumerId, consumerName);
     const l = rows.length;
     const args: (Filter["value"])[] = rows.map(row => row.id); // Extract the 'id' field from each row;
-    let query = `UPDATE "${table}" SET `;
+    let query = `UPDATE "${quoteIfUppercase(table)}" SET `;
     let i = args.length+1;
     for (const p of this._props) {
       if (rows[0][p] === undefined) // do not create case if prop is not in the first row