+++ src/filter/filter.ts
@@ -1,11 +1,12 @@
 import { addConditions } from "./condition";
+import { quoteIfUppercase } from "../crud/quote";
 import type { Filters, Filter, Sort, LogicalOperator } from "../types";
 
 function filter(
   first: number,
   rows: number | null,
   sortField: string | null,
-  sortOrder: Sort,
+  sortOrder: Sort | null,
   filters: Filters | null,
 ): { filterClause: string, args: (Filter["value"])[] } {
   
@@ -30,10 +31,11 @@ function where(conditions: string[], operator: LogicalOperator = "AND"): string 
 }
 
 // Adds order by clause
-function orderBy(sortField: string | null, sortOrder: Sort): string {
+function orderBy(sortField: string | null, sortOrder: Sort | null): string {
   if (!sortField) 
     return "";
-  return ` ORDER BY "${sortField}" ${sortOrder}`;
+  const o = sortOrder || "ASC";
+  return ` ORDER BY ${quoteIfUppercase(sortField)} ${o}`;
 }
 
 // Adds limit clause