+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.1.1 (Sep 10th 2025)
+# 0.3.4 (Sep 10th 2025)
   
 - Update dependencies : 
   - "@dwtechs/checkard": "3.5.1",
@@ -6,6 +6,43 @@
   - "@dwtechs/sparray": "0.2.1",
 
 
-# 0.1.0 (Dec 28th 2024)
+# 0.3.3 (Aug 16th 2025)
+
+- Improved error handling in entity operations
+- Update Antity.js to version 0.11.1
+
+
+# 0.3.2 (May 2nd 2025)
+
+- Upgrade @dwtechs/antity dependency to 0.10.0
+
+
+# 0.3.1 (May 1st 2025)
+
+- Upgrade @dwtechs/antity dependency to 0.9.2
+
+
+# 0.3.0 (Apr 27th 2025)
+
+- Enhance quote handling for reserved keywords
+
+
+# 0.2.1 (Apr 26th 2025)
+
+- Add proper quote handling for table names
+
+
+# 0.2.0 (Apr 25th 2025)
+
+- Add execute() method to the library
+- Add quotes around property names with uppercases
+
+
+# 0.1.1 (Apr 24th 2025)
+
+- fix package.json file
+
+
+# 0.1.0 (Apr 23th 2025)
 
 - initial release